+++ src/ui/control/geolocate_control.test.ts
@@ -111,8 +111,116 @@ describe('GeolocateControl with no options', () => {
         expect(geolocate._geolocateButton.disabled).toBeFalsy();
     });
 
-    test('error event', async () => {
+    test('error event in waiting active state', async () => {
         const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        geolocate._watchState = 'WAITING_ACTIVE';
+
+        const click = new window.Event('click');
+        const errorPromise = geolocate.once('error');
+        geolocate._geolocateButton.dispatchEvent(click);
+
+        geolocation.sendError({code: 2, message: 'error message'});
+        const error = await errorPromise;
+
+        expect(error.code).toBe(2);
+        expect(error.message).toBe('error message');
+        expect(geolocate._watchState).toBe('ACTIVE_ERROR');
+    });
+
+    test('error event in active lock state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        geolocate._watchState = 'ACTIVE_LOCK';
+
+        const click = new window.Event('click');
+        const errorPromise = geolocate.once('error');
+        geolocate._geolocateButton.dispatchEvent(click);
+
+        geolocation.sendError({code: 2, message: 'error message'});
+        const error = await errorPromise;
+
+        expect(error.code).toBe(2);
+        expect(error.message).toBe('error message');
+        expect(geolocate._watchState).toBe('ACTIVE_ERROR');
+    });
+
+    test('error event in background state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        geolocate._watchState = 'BACKGROUND';
+
+        const click = new window.Event('click');
+        const errorPromise = geolocate.once('error');
+        geolocate._geolocateButton.dispatchEvent(click);
+
+        geolocation.sendError({code: 2, message: 'error message'});
+        const error = await errorPromise;
+
+        expect(error.code).toBe(2);
+        expect(error.message).toBe('error message');
+        expect(geolocate._watchState).toBe('BACKGROUND_ERROR');
+    });
+
+    test('error event in active error state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        geolocate._watchState = 'ACTIVE_ERROR';
+
+        const click = new window.Event('click');
+        const errorPromise = geolocate.once('error');
+        geolocate._geolocateButton.dispatchEvent(click);
+
+        geolocation.sendError({code: 2, message: 'error message'});
+        const error = await errorPromise;
+
+        expect(error.code).toBe(2);
+        expect(error.message).toBe('error message');
+        expect(geolocate._watchState).toBe('ACTIVE_ERROR');
+    });
+
+    test('error event in background error state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        geolocate._watchState = 'BACKGROUND_ERROR';
+
+        const click = new window.Event('click');
+        const errorPromise = geolocate.once('error');
+        geolocate._geolocateButton.dispatchEvent(click);
+
+        geolocation.sendError({code: 2, message: 'error message'});
+        const error = await errorPromise;
+
+        expect(error.code).toBe(2);
+        expect(error.message).toBe('error message');
+        expect(geolocate._watchState).toBe('BACKGROUND_ERROR');
+    });
+
+    test('error event in off state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        geolocate._watchState = 'OFF';
+
+        const click = new window.Event('click');
+        const errorPromise = geolocate.once('error');
+        geolocate._geolocateButton.dispatchEvent(click);
+
+        geolocation.sendError({code: 2, message: 'error message'});
+        const error = await errorPromise;
+
+        expect(error.code).toBe(2);
+        expect(error.message).toBe('error message');
+        expect(geolocate._watchState).toBe('OFF');
+    });
+
+    test('error event when trackUserLocation is false', async () => {
+        const geolocate = new GeolocateControl({trackUserLocation: false});
         map.addControl(geolocate);
         await sleep(0);
         const click = new window.Event('click');
@@ -124,6 +232,7 @@ describe('GeolocateControl with no options', () => {
 
         expect(error.code).toBe(2);
         expect(error.message).toBe('error message');
+        expect(geolocate._watchState).toBeUndefined();
     });
 
     test('does not throw if removed quickly', () => {
@@ -135,6 +244,26 @@ describe('GeolocateControl with no options', () => {
         const geolocate = new GeolocateControl(undefined);
         map.addControl(geolocate);
         map.removeControl(geolocate);
+    });
+
+    test('outofmaxbounds event in waiting active state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        map.setMaxBounds([[0, 0], [10, 10]]);
+        geolocate._watchState = 'WAITING_ACTIVE';
+
+        const click = new window.Event('click');
+
+        const promise = geolocate.once('outofmaxbounds');
+        geolocate._geolocateButton.dispatchEvent(click);
+        geolocation.send({latitude: 10, longitude: 20, accuracy: 3, timestamp: 4});
+        const position = await promise;
+        expect(geolocate._watchState).toBe('ACTIVE_ERROR');
+        expect(position.coords.latitude).toBe(10);
+        expect(position.coords.longitude).toBe(20);
+        expect(position.coords.accuracy).toBe(3);
+        expect(position.timestamp).toBe(4);
     });
 
     test('outofmaxbounds event in active lock state', async () => {
@@ -172,6 +301,85 @@ describe('GeolocateControl with no options', () => {
         geolocation.send({latitude: 10, longitude: 20, accuracy: 3, timestamp: 4});
         const position = await promise;
         expect(geolocate._watchState).toBe('BACKGROUND_ERROR');
+        expect(position.coords.latitude).toBe(10);
+        expect(position.coords.longitude).toBe(20);
+        expect(position.coords.accuracy).toBe(3);
+        expect(position.timestamp).toBe(4);
+    });
+
+    test('outofmaxbounds event in active error state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        map.setMaxBounds([[0, 0], [10, 10]]);
+        geolocate._watchState = 'ACTIVE_ERROR';
+
+        const click = new window.Event('click');
+
+        const promise = geolocate.once('outofmaxbounds');
+        geolocate._geolocateButton.dispatchEvent(click);
+        geolocation.send({latitude: 10, longitude: 20, accuracy: 3, timestamp: 4});
+        const position = await promise;
+        expect(geolocate._watchState).toBe('ACTIVE_ERROR');
+        expect(position.coords.latitude).toBe(10);
+        expect(position.coords.longitude).toBe(20);
+        expect(position.coords.accuracy).toBe(3);
+        expect(position.timestamp).toBe(4);
+    });
+
+    test('outofmaxbounds event in background error state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        map.setMaxBounds([[0, 0], [10, 10]]);
+        geolocate._watchState = 'BACKGROUND_ERROR';
+
+        const click = new window.Event('click');
+
+        const promise = geolocate.once('outofmaxbounds');
+        geolocate._geolocateButton.dispatchEvent(click);
+        geolocation.send({latitude: 10, longitude: 20, accuracy: 3, timestamp: 4});
+        const position = await promise;
+        expect(geolocate._watchState).toBe('BACKGROUND_ERROR');
+        expect(position.coords.latitude).toBe(10);
+        expect(position.coords.longitude).toBe(20);
+        expect(position.coords.accuracy).toBe(3);
+        expect(position.timestamp).toBe(4);
+    });
+
+    test('outofmaxbounds event in off state', async () => {
+        const geolocate = new GeolocateControl(undefined);
+        map.addControl(geolocate);
+        await sleep(0);
+        map.setMaxBounds([[0, 0], [10, 10]]);
+        geolocate._watchState = 'OFF';
+
+        const click = new window.Event('click');
+
+        const promise = geolocate.once('outofmaxbounds');
+        geolocate._geolocateButton.dispatchEvent(click);
+        geolocation.send({latitude: 10, longitude: 20, accuracy: 3, timestamp: 4});
+        const position = await promise;
+        expect(geolocate._watchState).toBe('OFF');
+        expect(position.coords.latitude).toBe(10);
+        expect(position.coords.longitude).toBe(20);
+        expect(position.coords.accuracy).toBe(3);
+        expect(position.timestamp).toBe(4);
+    });
+
+    test('outofmaxbounds event when trackUserLocation = false', async () => {
+        const geolocate = new GeolocateControl({trackUserLocation: false});
+        map.addControl(geolocate);
+        await sleep(0);
+        map.setMaxBounds([[0, 0], [10, 10]]);
+
+        const click = new window.Event('click');
+
+        const promise = geolocate.once('outofmaxbounds');
+        geolocate._geolocateButton.dispatchEvent(click);
+        geolocation.send({latitude: 10, longitude: 20, accuracy: 3, timestamp: 4});
+        const position = await promise;
+        expect(geolocate._watchState).toBeUndefined();
         expect(position.coords.latitude).toBe(10);
         expect(position.coords.longitude).toBe(20);
         expect(position.coords.accuracy).toBe(3);